Fix offline reload
